Removing images

+++ database/migrations/2019_07_07_123900_create_courses_table.php
@@ -14,14 +14,14 @@ class CreateCoursesTable extends Migration
     public function up()
     {
         Schema::create('courses', function (Blueprint $table) {
-            $table->bigIncrements('id');
+            $table->increments('id');
             $table->string('name');
             $table->longText('description');
             $table->string('logo');
             $table->string('schedule_integrated')->nullable();
             $table->string('schedule_subsequent')->nullable();
 
-            $table->unsignedBigInteger('user_id');
+            $table->unsignedInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
 
             $table->timestamps();
